replace api.js by api.ts
